Location bugs fixed

+++ app/src/main/java/com/example/myapplication/data/repository/restaurant/RestaurantEmployeesRepository.java
@@ -9,6 +9,7 @@ import static com.example.myapplication.presentation.utils.Utils.USER_NAME_KEY;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.ACTIVE_ORDERS_COUNT;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.COOK;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.COOKS;
+import static com.example.myapplication.presentation.utils.constants.Restaurant.COOKS_COUNT;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.COOK_QR_CODE;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.IS_WORKING;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.LOCATION;
@@ -17,12 +18,14 @@ import static com.example.myapplication.presentation.utils.constants.Restaurant.
 import static com.example.myapplication.presentation.utils.constants.Restaurant.RESTAURANT_LOCATION;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.WAITER;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.WAITERS;
+import static com.example.myapplication.presentation.utils.constants.Restaurant.WAITERS_COUNT;
 import static com.example.myapplication.presentation.utils.constants.Restaurant.WAITER_QR_CODE;
 
 import android.net.Uri;
 
 import com.example.myapplication.data.dto.common.EmployeeDto;
 import com.example.myapplication.data.dto.restaurant.EmployeeRestaurantDto;
+import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 
@@ -38,7 +41,7 @@ import io.reactivex.rxjava3.core.SingleEmitter;
 
 public class RestaurantEmployeesRepository {
 
-    public Single<Boolean> haveOrders(String restaurantId, String locationId){
+    public Single<Boolean> haveOrders(String restaurantId, String locationId) {
         return Single.create(emitter -> {
             service.fireStore
                     .collection(RESTAURANT_LIST)
@@ -48,10 +51,10 @@ public class RestaurantEmployeesRepository {
                     .collection(WAITERS)
                     .document(service.auth.getCurrentUser().getUid())
                     .get().addOnCompleteListener(task -> {
-                        if (task.isSuccessful()){
+                        if (task.isSuccessful()) {
                             List<String> orders = (List<String>) task.getResult().get(ACTIVE_ORDERS_COUNT);
                             assert orders != null;
-                            if (!orders.isEmpty()){
+                            if (!orders.isEmpty()) {
                                 emitter.onSuccess(true);
                             } else {
                                 emitter.onSuccess(false);
@@ -63,28 +66,43 @@ public class RestaurantEmployeesRepository {
         });
     }
 
-    public Completable deleteRestaurantEmployee(String restaurantId, String locationId, String userId, String role){
+    public Completable deleteRestaurantEmployee(String restaurantId, String locationId, String userId, String role) {
         return Completable.create(emitter -> {
-            String document;
-            if (role.equals(WAITER)){
-                document = WAITERS;
-            } else {
-                document = COOKS;
-            }
-            service.fireStore
+            DocumentReference locationRef = service.fireStore
                     .collection(RESTAURANT_LIST)
                     .document(restaurantId)
                     .collection(RESTAURANT_LOCATION)
-                    .document(locationId)
-                    .collection(document)
-                    .document(userId)
-                    .delete().addOnCompleteListener(task -> {
-                        if (task.isSuccessful()){
-                            emitter.onComplete();
-                        } else {
-                            emitter.onError(new Throwable());
+                    .document(locationId);
+            String document;
+            String key;
+            if (role.equals(WAITER)) {
+                key = WAITERS_COUNT;
+                document = WAITERS;
+            } else {
+                key = COOKS_COUNT;
+                document = COOKS;
+            }
+
+            locationRef.get().addOnCompleteListener(taskGet -> {
+                if (taskGet.isSuccessful()) {
+                    int currentCount = Integer.parseInt(taskGet.getResult().getString(key));
+                    locationRef.update(key, String.valueOf(currentCount + 1)).addOnCompleteListener(taskUpdate -> {
+                        if (taskUpdate.isSuccessful()) {
+                            locationRef
+                                    .collection(document)
+                                    .document(userId)
+                                    .delete().addOnCompleteListener(task -> {
+                                        if (task.isSuccessful()) {
+
+                                            emitter.onComplete();
+                                        } else {
+                                            emitter.onError(new Throwable());
+                                        }
+                                    });
                         }
                     });
+                }
+            });
         });
     }
 
@@ -99,6 +117,7 @@ public class RestaurantEmployeesRepository {
             List<EmployeeRestaurantDto> employees = new ArrayList<>();
             getCooks(employees, docRef, emitter);
             getWaiters(employees, docRef, emitter);
+            emitter.onSuccess(employees);
         });
     }
 
@@ -148,6 +167,8 @@ public class RestaurantEmployeesRepository {
                     .collection(waiterPath)
                     .document(userId);
 
+            DocumentReference locationRef = service.fireStore.collection(waiterPath).getParent();
+
             service.fireStore.collection(USER_LIST).document(userId)
                     .get().addOnCompleteListener(task -> {
                         if (task.isSuccessful()) {
@@ -157,13 +178,25 @@ public class RestaurantEmployeesRepository {
                             waiter.put(IS_WORKING, false);
                             waiter.put(ACTIVE_ORDERS_COUNT, Collections.emptyList());
 
-                            docRef.set(waiter).addOnCompleteListener(taskWaiter -> {
-                                if (taskWaiter.isSuccessful()) {
-                                    emitter.onComplete();
-                                } else {
-                                    emitter.onError(new Throwable(taskWaiter.getException()));
+                            locationRef.get().addOnCompleteListener(taskGet -> {
+                                if (taskGet.isSuccessful()) {
+
+                                    int currentCount = Integer.parseInt(taskGet.getResult().getString(WAITERS_COUNT));
+                                    locationRef.update(WAITERS_COUNT, String.valueOf(currentCount + 1)).addOnCompleteListener(taskUpdate -> {
+                                        if (taskUpdate.isSuccessful()) {
+                                            docRef.set(waiter).addOnCompleteListener(taskWaiter -> {
+                                                if (taskWaiter.isSuccessful()) {
+                                                    emitter.onComplete();
+                                                } else {
+                                                    emitter.onError(new Throwable(taskWaiter.getException()));
+                                                }
+                                            });
+                                        }
+                                    });
                                 }
                             });
+
+
                         }
                     });
         });
@@ -175,6 +208,8 @@ public class RestaurantEmployeesRepository {
                     .collection(path)
                     .document(service.auth.getCurrentUser().getUid());
 
+            DocumentReference locationRef = service.fireStore.collection(path).getParent();
+
             service.fireStore.collection(USER_LIST).document(service.auth.getCurrentUser().getUid())
                     .get().addOnCompleteListener(task -> {
                         if (task.isSuccessful()) {
@@ -182,11 +217,25 @@ public class RestaurantEmployeesRepository {
                             HashMap<String, String> cook = new HashMap<>();
                             cook.put(USER_NAME_KEY, name);
 
-                            docRef.set(cook).addOnCompleteListener(taskAdd -> {
-                                if (taskAdd.isSuccessful()) {
-                                    emitter.onComplete();
+                            locationRef.get().addOnCompleteListener(taskGet -> {
+                                if (taskGet.isSuccessful()) {
+
+                                    int currentCount = Integer.parseInt(taskGet.getResult().getString(COOKS_COUNT));
+                                    locationRef.update(COOKS_COUNT, String.valueOf(currentCount + 1)).addOnCompleteListener(taskUpdate -> {
+                                        if (taskUpdate.isSuccessful()) {
+                                            docRef.set(cook).addOnCompleteListener(taskAdd -> {
+                                                if (taskAdd.isSuccessful()) {
+                                                    emitter.onComplete();
+                                                } else {
+                                                    emitter.onError(new Throwable(taskAdd.getException()));
+                                                }
+                                            });
+                                        } else {
+                                            emitter.onError(new Throwable(taskUpdate.getException()));
+                                        }
+                                    });
                                 } else {
-                                    emitter.onError(new Throwable(taskAdd.getException()));
+                                    emitter.onError(new Throwable(taskGet.getException()));
                                 }
                             });
                         } else {
@@ -304,7 +353,7 @@ public class RestaurantEmployeesRepository {
 
     private void getWaiters(List<EmployeeRestaurantDto> employees, DocumentReference docRef, SingleEmitter<List<EmployeeRestaurantDto>> emitter) {
         docRef.collection(WAITERS).get().addOnCompleteListener(task -> {
-            if (task.isSuccessful()){
+            if (task.isSuccessful()) {
                 List<DocumentSnapshot> snapshots = task.getResult().getDocuments();
                 for (DocumentSnapshot current : snapshots) {
                     employees.add(new EmployeeRestaurantDto(
